+++ packages/core/src/element.ts
@@ -57,11 +57,10 @@ export default class ImpulseElement extends HTMLElement {
 
   disconnectedCallback() {
     // Order is important
+    this.disconnected();
     this.action.stop();
     this.target.stop();
     this.targets.stop();
-    // We want to invoke the `disconnected` callback after stopping the `target(s)` but before stopping the `property`
-    this.disconnected();
     this.property.stop();
     this._started = false;
   }