+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Export `AttributeObserver` and `ElementObserver`
 - Support number values with underscores
 - Add `lazyImport` function which imports the elements/targets lazily